feat(authentications): make GPA a DecimalField
this allows us to set the decimal places in the field instead of in a string format, eliminating confusion when some writes e.g. 4.500 in their profile but profile says 4.5 after save

+++ tjdests/apps/authentication/models.py
@@ -8,7 +8,7 @@ class User(AbstractUser):
     accepted_terms = models.BooleanField(default=False)
     graduation_year = models.PositiveSmallIntegerField(null=True)
 
-    GPA = models.FloatField(null=True, blank=True, name="GPA", help_text="Weighted GPA")
+    GPA = models.DecimalField(null=True, blank=True, name="GPA", help_text="Weighted GPA", max_digits=4, decimal_places=3)
 
     is_senior = models.BooleanField(default=False)
     is_student = models.BooleanField(default=False)
